Refactored tests into unit and CI integration tests
