+++ ci/integration_tests/embedded_server/sshd_config.tmpl
